Title: 그 다음 알파벳, Time: 111ms, Memory: 10MB, Status: Passed - CodeTree

+++ 250115/그 다음 알파벳/next-alphabet.java
@@ -8,6 +8,6 @@ public class Main {
         if (c == 'z')
             System.out.print('a');
         else 
-            System.out.print((char)((int)c + 1));
+            System.out.print((char)(c + 1));
     }
 }
